style(db): split images
