+++ backend/matcha/services/pictures.py
@@ -7,9 +7,9 @@ from werkzeug.utils import secure_filename
 
 from matcha.app_utils import check_request_json, make_unique
 
-from matcha.db.db import (
-    db_upload_pictures,
+from matcha.db.images import (
     db_get_user_images,
+    db_upload_pictures,
     db_set_profile_picture,
     db_get_url_profile,
     db_count_number_image,
