+++ backend/matcha/services/profile.py
@@ -7,13 +7,13 @@ from flask_jwt_extended import (
 
 from matcha.app_utils import check_request_json
 
+from matcha.db.images import db_get_user_images, db_get_url_profile
+
 from matcha.db.db import (
     db_get_interests,
-    db_get_user_images,
     db_set_user_profile_data,
     db_get_user_per_id,
     db_get_user_per_username,
-    db_get_url_profile,
 )
 
 from matcha.db.like import db_get_is_liked, db_get_list_liked_by
